requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==3.4.4
defusedxml==0.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==0.5.0
django-allauth==0.44.0
django-ckeditor==6.0.0
django-environ==0.4.5
django-heroku==0.3.1
django-js-asset==1.2.2
django==3.1.6
djangocms-admin-style==2.0.2
gunicorn==20.0.4
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
numpy==1.20.1; python_version >= '3.7'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
opencv-python==4.5.1.48
pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt[crypto]==2.0.1; python_version >= '3.6'
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
shell==1.0.1
sqlparse==0.4.1; python_version >= '3.5'
urllib3==1.26.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
whitenoise==5.2.0
